% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsps_fit.R
\name{dsps_fit}
\alias{dsps_fit}
\title{Fitting (Training) dynamic survival prediction}
\usage{
dsps_fit(data, kernel, C, sigma = NULL)
}
\arguments{
\item{data}{a data frame with time-dependent variables at unique event time points.
\code{time}, \code{event}, \code{subjectid} should be included after the leading columns of time-dependent variables (Order doesn't matter).
\code{time} is the ordered unique event time points. \code{event} is the survival status at \code{time} (1:event; 0:non-event). \code{subjectid} is the unique identifier for subject.}

\item{kernel}{the kernel to be used; Four types of kernels are currently supported: "linear", "gaussian", "semi_linear", "semi_gaussian"}

\item{C}{a positive number, the penalty parameter that controls the penalty of mis-classification.}

\item{sigma}{a positive number, the bandwidth parameter for "gaussian" and "semi_gaussian" kernel.}
}
\value{
A list of fitting results that can be passed to function \code{dsp_predict}.
}
\description{
\code{dsp_fit} is used to fit the dynamic survival prediction.
It allows for both linear and Gaussian kernel as well as two time-discrete kernel, semi-linear and semi-Gaussian kernel as described in the manuscript.
A list of estimated parameters are returned.
}
\examples{

data_list <- sim_data(n=100, er=0.6)

# Keep only at-risk observations
train_data <- data_list$Z \%>\%
                filter(at_risk == 1) \%>\%
                select(-at_risk)

fit <- dsps_fit(train_data,
               kernel="semi_linear",
               C=2^0)

fit <- dsps_fit(train_data,
               kernel="semi_gaussian",
               sigma=2^0,
               C=2^0)

}
